+++ apps/carp_mobile_sensing_app/lib/src/app.dart
@@ -21,12 +21,6 @@ class LoadingPage extends StatelessWidget {
   ///
   /// Returns true when successfully done.
   Future<bool> init(BuildContext context) async {
-    // Request location "always" permissions upfront.
-    // Note that this is a two-step process on Android, where the user first has
-    // to grant "when in use" permissions, and then "always" permissions.
-    await Permission.locationWhenInUse.request();
-    await Permission.locationAlways.request();
-
     // Initialize and use the CAWS backend if not in local deployment mode
     if (bloc.deploymentMode != DeploymentMode.local) {
       await CarpBackend().initialize();