+++ backends/carp_webservices/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.0.0
+
+* require `carp_mobile_sensing` ^3.0.0
+
 ## 4.2.0
 
 * require `carp_serializable` ^3.0.0, which replaces the built-in `Uuid` with the [uuid](https://pub.dev/packages/uuid) package